Create Task_6
This web form is for ordering computer parts. The fields for CPUs, GPUs and RAM are dynamic. It was made using HTML, CSS and JavaScript. Real-time validation of forms, component addition/removals and a user-friendly design were also included. This is great for learning or adding to an e-commerce project.